move fetchExperimentalPatches from AilaChat
